+++ src/tab/TabPosition.java
@@ -104,7 +104,7 @@ public class TabPosition implements Comparable<TabPosition>, Copyable<TabPositio
 	 * @param divisor The amount to divide up the units of a whole note.<br>
 	 * 	i.e. use 4 to quantize to quarter notes, use 6 to quantize to dotted quarter notes, etc
 	 */
-	public void quantize(TimeSignature sig, int divisor){
+	public void quantize(TimeSignature sig, double divisor){
 		this.getPosition().quantize(sig, divisor);
 	}
